requirements.txt
requests
numpy
pandas
tensorflow-cpu
scikit-learn